requirements.txt
blinker==1.9.0
braintree==4.24.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
Flask==2.3.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.0.1
requests==2.32.5
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
